refactor: extract clock menu
