+++ src/main/java/com/holub/ui/UniversePanel.java
@@ -6,6 +6,7 @@ import com.holub.tools.Observer;
 
 import com.holub.ui.cell.CellUI;
 import com.holub.ui.cell.CellUIFactory;
+import com.holub.ui.menu.MenuSite;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Graphics;
