+++ src/main/java/com/holub/ui/menu/MenuSite.java
@@ -1,4 +1,4 @@
-package com.holub.ui;
+package com.holub.ui.menu;
 
 
 import java.awt.Component;
