+++ src/test/java/com/holub/ui/menu/MenuSiteTest.java
@@ -1,13 +1,10 @@
-package com.holub.ui;
-
-import static org.junit.jupiter.api.Assertions.*;
+package com.holub.ui.menu;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.file.Path;
 import java.nio.file.Paths;
